Remove arbitrary password length limit.
The password length is currently limited to 50 characters but longer passwords are possible.  If passwords longer than a certain length shouldn't be specified this should be checked and an error issued once the data is submitted otherwise what happens is the password is truncated but the user is not aware until they go to use the password and it doesn't work.